requirements.txt
Flask == 3.1.0
tabula-py == 2.10.0
pandas == 2.2.3
jpype1 == 1.5.2